% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explor.R, R/explor_MCA.R, R/explor_PCA.R
\name{explor}
\alias{explor}
\alias{explor.MCA}
\alias{explor.PCA}
\title{Interface for analysis results exploration}
\usage{
explor(obj)

\method{explor}{MCA}(obj)

\method{explor}{PCA}(obj)
}
\arguments{
\item{obj}{object containing analysis results}
}
\value{
The function launches a shiny app in the system web browser.
}
\description{
This function launches a shiny app in a web browser in order to do
interactive visualisation and exploration of an analysis results.
}
\examples{
\dontrun{

require(FactoMineR)

## MCA exploration
data(hobbies)
mca <- MCA(hobbies[1:1000,c(1:8,21:23)], quali.sup = 9:10, 
           quanti.sup = 11, ind.sup = 1:100, graph = FALSE)
explor(mca)

## PCA exploration
data(decathlon)
d <- decathlon[,1:12]
pca <- PCA(d, quanti.sup = 11:12, graph = FALSE)
explor(pca)
}
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>
}

