% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggvar}
\alias{ggvar}
\alias{ggvar.MCA}
\title{Graphical representation of the variables (columnss) of a multivariate analysis}
\usage{
ggvar(obj, ...)

\method{ggvar}{MCA}(obj, xax = 1, yax = 2, size = 4, alpha = 0.5, palette = "Set1", ...)
}
\arguments{
\item{obj}{a multivariate analysis results object. Currently only MCA is supported}

\item{...}{arguments passed to other methods}

\item{xax}{number of the x axis}

\item{yax}{number of the y axis}

\item{size}{text size}

\item{alpha}{points opacity}

\item{palette}{palette for variables coloring}
}
\description{
This function displays a graphical representation of the variables (columns) of a multivariate analysis.

This function displays a graphical representation of the variables
(columns) of a multiple correspondence analysis generated by the \code{MCA}
function of the \code{FactoMineR} package.
}
\seealso{
\code{\link[FactoMineR]{MCA}}
}
