% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imca.R
\name{imca}
\alias{imca}
\title{Interface for multiple correspondence analysis}
\usage{
imca(acm)
}
\arguments{
\item{acm}{an object of class MCA, result of the \code{MCA()} function from the \code{FactoMineR} package.}
}
\value{
The function launches a shiny app in the system web browser.
}
\description{
This function launches a shiny app in a web browser in order to do
interactive visualisation and exploration of a multiple correspondence analysis.
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>
}
\seealso{
\code{\link[FactoMineR]{MCA}}
}

