% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA_plots.R
\name{MCA_biplot}
\alias{MCA_biplot}
\title{Interactive MCA biplot}
\usage{
MCA_biplot(res, xax = 1, yax = 2, col_var, ind_sup = TRUE,
  var_sup = TRUE, bi_lab_min_contrib = 0, symbol_var = NULL,
  ind_point_size = 16, var_point_size = 96, ind_opacity = 0.5,
  ind_opacity_var = NULL, ind_labels = FALSE, zoom_callback = NULL,
  in_explor = FALSE, ...)
}
\arguments{
\item{res}{Result of prepare_results() call}

\item{xax}{Horizontal axis number}

\item{yax}{Vertical axis number}

\item{col_var}{name of the variable for points color}

\item{ind_sup}{TRUE to display supplementary individuals}

\item{var_sup}{TRUE to display supplementary variables}

\item{bi_lab_min_contrib}{Contribution threshold to display points labels}

\item{symbol_var}{name of the variable for points symbol}

\item{ind_point_size}{base point size for individuals}

\item{var_point_size}{base point size for variable levels}

\item{ind_opacity}{individuals point opacity (constant)}

\item{ind_opacity_var}{individuals point opacity (variable)}

\item{ind_labels}{TRUE to display individuals labels}

\item{zoom_callback}{scatterD3 zoom callback JavaScript body}

\item{in_explor}{wether the plot is to be displayed in the \code{explor} interface}

\item{...}{Other arguments passed to scatterD3}
}
\description{
This function generates an HTML widget displaying the variables plot of an MCA result.
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>
}
