% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA_plots.R
\name{MCA_ind_plot}
\alias{MCA_ind_plot}
\title{Interactive MCA indivuals plot}
\usage{
MCA_ind_plot(res, xax = 1, yax = 2, ind_sup = TRUE, ind_col = NULL,
  symbol_var = "Type", size_var = NULL, size_range = c(10, 400),
  lab_var = "Name", ellipses = FALSE, ind_point_size = 64,
  ind_labels_size = 9, ind_opacity = 0.5, transitions = FALSE)
}
\arguments{
\item{res}{Result of prepare_results() call}

\item{xax}{Horizontal axis number}

\item{yax}{Vertical axis number}

\item{ind_sup}{TRUE to display supplementary individuals}

\item{ind_col}{variable to be used for points color}

\item{symbol_var}{name of the variable for points symbol}

\item{size_var}{name of the variable for points size}

\item{size_range}{points size range with format c(minimum, maximum)}

\item{lab_var}{variable to be used for points names}

\item{ellipses}{wether to display confidence ellipses for points color}

\item{ind_point_size}{fixed points size}

\item{ind_labels_size}{points labels size}

\item{ind_opacity}{points opacity}

\item{transitions}{wether to display animated transitions (only when used in a shiny app)}
}
\description{
This function generates an HTML widget displaying the individuals plot of an MCA result.
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>
}

