% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca_prepare.R
\name{mca_prepare.MCA}
\alias{mca_prepare.MCA}
\title{Results preparation for a FactoMineR MCA analysis}
\usage{
mca_prepare.MCA(mca)
}
\arguments{
\item{acm}{an object of class MCA, result of the \code{MCA()} function from the \code{FactoMineR} package.}
}
\description{
This function prepares MCA results to be used by \code{imca}. Not to be used directly.
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>
}
\seealso{
\code{\link[FactoMineR]{MCA}}
}

