% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA_plots.R
\name{MCA_var_plot}
\alias{MCA_var_plot}
\title{Interactive MCA variables plot}
\usage{
MCA_var_plot(res, xax = 1, yax = 2, var_sup = TRUE,
  var_lab_min_contrib = 0, col_var = "Variable", symbol_var = NULL,
  size_var = NULL, size_range = c(10, 400), var_lab_size = 10,
  var_point_size = 30, transitions = FALSE)
}
\arguments{
\item{res}{Result of prepare_results() call}

\item{xax}{Horizontal axis number}

\item{yax}{Vertical axis number}

\item{var_sup}{TRUE to display supplementary variables}

\item{var_lab_min_contrib}{Contribution threshold to display points labels}

\item{col_var}{name of the variable for points color}

\item{symbol_var}{name of the variable for points symbol}

\item{size_var}{name of the variable for points size}

\item{size_range}{points size range with format c(minimum, maximum)}

\item{var_lab_size}{points label size}

\item{var_point_size}{fixed points size}

\item{transitions}{wether to display animated transitions (only when used in a shiny app)}
}
\description{
This function generates an HTML widget displaying the variables plot of an MCA result.
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>
}

