% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plots.R
\name{PCA_var_plot}
\alias{PCA_var_plot}
\title{Interactive PCA variables plot}
\usage{
PCA_var_plot(res, xax = 1, yax = 2, var_sup = TRUE,
  var_lab_min_contrib = 0, scale_unit = FALSE, col_var = NULL,
  size_var = NULL, zoom_callback = NULL, in_explor = FALSE, xlim = NULL,
  ylim = NULL, ...)
}
\arguments{
\item{res}{Result of prepare_results() call}

\item{xax}{Horizontal axis number}

\item{yax}{Vertical axis number}

\item{var_sup}{TRUE to display supplementary variables}

\item{var_lab_min_contrib}{Contribution threshold to display points labels}

\item{scale_unit}{wether the PCA is scaled}

\item{col_var}{name of the variable for points color}

\item{size_var}{name of the variable for points size}

\item{zoom_callback}{scatterD3 zoom callback JavaScript body}

\item{in_explor}{wether the plot is to be displayed in the \code{explor} interface}

\item{xlim}{custom x axis limits}

\item{ylim}{custom y axis limits}

\item{...}{Other arguments passed to scatterD3}
}
\description{
This function generates an HTML widget displaying the variables plot of a PCA result.
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>
}

