% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_results.R, R/prepare_results_MCA.R, R/prepare_results_PCA.R, R/prepare_results_dudi_mca.R
\name{prepare_results}
\alias{prepare_results}
\alias{prepare_results.MCA}
\alias{prepare_results.PCA}
\alias{prepare_results.acm}
\title{Analysis results preparation}
\usage{
prepare_results(obj)

\method{prepare_results}{MCA}(obj)

\method{prepare_results}{PCA}(obj)

\method{prepare_results}{acm}(obj)
}
\arguments{
\item{obj}{object containing analysis results}
}
\description{
This function prepares results to be used by \code{explor}. Not to be used directly.
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>

Julien Barnier <julien.barnier@ens-lyon.fr>

Julien Barnier <julien.barnier@ens-lyon.fr>

Julien Barnier <julien.barnier@ens-lyon.fr>
}
\seealso{
\code{\link[FactoMineR]{MCA}}

\code{\link[FactoMineR]{PCA}}

\code{\link[ade4]{dudi.acm}}
}

