% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{speMCA_varsup}
\alias{speMCA_varsup}
\title{Compute supplementary variables data for a GDAtools::speMCA result}
\usage{
speMCA_varsup(mca, df)
}
\arguments{
\item{mca}{result object from speMCA.}

\item{df}{data frame with the supplementary variables data. Must have the
same number of rows than the data used with speMCA.}
}
\value{
A list of results suitable to be added as a `supv` element to the `mca`
object.
}
\description{
Compute supplementary variables data for a GDAtools::speMCA result
}
\seealso{
\code{\link[GDAtools]{speMCA}}, \code{\link[GDAtools]{varsup}}
}
