% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explor.R, R/explor_MCA.R
\name{explor}
\alias{explor}
\alias{explor.MCA}
\title{Interface for analysis results exploration}
\usage{
explor(obj)

\method{explor}{MCA}(obj)
}
\arguments{
\item{obj}{object containing analysis results}
}
\value{
The function launches a shiny app in the system web browser.

The function launches a shiny app in the system web browser.
}
\description{
This function launches a shiny app in a web browser in order to do
interactive visualisation and exploration of an analysis results.
}
\details{
Interface for multiple correspondence analysis

This function launches a shiny app in a web browser in order to do
interactive visualisation and exploration of a multiple correspondence analysis.
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>

Julien Barnier <julien.barnier@ens-lyon.fr>
}
\seealso{
\code{\link[FactoMineR]{MCA}}
}

