% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA_plots.R
\name{MCA_ind_plot}
\alias{MCA_ind_plot}
\title{Interactive MCA indivuals plot}
\usage{
MCA_ind_plot(
  res,
  xax = 1,
  yax = 2,
  ind_sup = TRUE,
  ind_lab_min_contrib = 0,
  lab_var = NULL,
  col_var = NULL,
  symbol_var = NULL,
  opacity_var = NULL,
  size_var = NULL,
  size_range = c(10, 300),
  zoom_callback = NULL,
  in_explor = FALSE,
  ...
)
}
\arguments{
\item{res}{Result of prepare_results() call}

\item{xax}{Horizontal axis number}

\item{yax}{Vertical axis number}

\item{ind_sup}{TRUE to display supplementary individuals}

\item{ind_lab_min_contrib}{Contribution threshold to display points labels}

\item{lab_var}{variable to be used for points names}

\item{col_var}{variable to be used for points color}

\item{symbol_var}{name of the variable for points symbol}

\item{opacity_var}{name of the variable for points opacity}

\item{size_var}{name of the variable for points size}

\item{size_range}{points size range with format c(minimum, maximum)}

\item{zoom_callback}{scatterD3 zoom callback JavaScript body}

\item{in_explor}{wether the plot is to be displayed in the \code{explor} interface}

\item{...}{Other arguments passed to scatterD3}
}
\description{
This function generates an HTML widget displaying the individuals plot of an MCA result.
}
