% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_results.R, R/prepare_results_MCA.R
\name{prepare_results}
\alias{prepare_results}
\alias{prepare_results.MCA}
\title{Results preparation for a FactoMineR MCA analysis}
\usage{
prepare_results(...)

\method{prepare_results}{MCA}(mca)
}
\arguments{
\item{...}{arguments passed to other methods}

\item{acm}{an object of class MCA, result of the \code{MCA()} function from the \code{FactoMineR} package.}
}
\description{
This function prepares MCA results to be used by \code{explor}. Not to be used directly.
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>

Julien Barnier <julien.barnier@ens-lyon.fr>
}
\seealso{
\code{\link[FactoMineR]{MCA}}
}

